add animation rig content validator
